push more abstractions into framework
